requirements.txt
absl-py==0.7.1
asn1crypto==0.24.0
astor==0.8.0
cryptography==2.1.4
cycler==0.10.0
decorator==4.4.2
gast==0.2.2
grpcio==1.22.0
h5py==2.9.0
idna==2.6
imageio==2.9.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
Markdown==3.1.1
matplotlib==3.3.4
mock==3.0.5
networkx==2.5.1
numpy==1.16.4
opencv-contrib-python==4.5.3.56
Pillow==8.3.2
protobuf==3.9.0
pycrypto==2.6.1
pygobject==3.26.1
pyparsing==2.4.7
python-apt==1.6.4
python-dateutil==2.8.2
PyWavelets==1.1.1
pyxdg==0.25
scikit-image==0.17.2
scipy==1.5.4
SecretStorage==2.3.1
six==1.11.0
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.2
termcolor==1.1.0
tifffile==2020.9.3
Werkzeug==0.15.4
